Started writing decode part
